refactor(test): convert to direct imports
