+++ test/__mocks__/api/notice-manager/createAnnContentResponse.ts
@@ -1,4 +1,4 @@
-import { ValueOf } from '@/types'
+import { ValueOf } from '@/types/types'
 import { NoticeLanguage } from '@/types/sg-hk4e-api'
 import type { APIGetAnnContent } from '@/types/sg-hk4e-api/response'
